move ROOT-aware code from walker.py into rootio.py and default object-reading into  class
